// src/utils/kaikas-types.ts
export interface Kaikas extends InjectedProviders {
  on?: (...args: any[]) => void
  removeListener?: (...args: any[]) => void
  providers?: Kaikas

  /**
   * EIP-747: Add wallet_watchAsset to Provider
   * https://eips.ethereum.org/EIPS/eip-747
   */
  request(args: { method: 'wallet_watchAsset'; params: WatchAssetParams }): Promise<boolean>

  /**
   * EIP-1193: Ethereum Provider JavaScript API
   * https://eips.ethereum.org/EIPS/eip-1193
   */
  request(
    args:
      | { method: 'klay_accounts' }
      | { method: 'klay_requestAccounts' }
      | { method: 'eth_accounts' }
      | { method: 'eth_requestAccounts' }
  ): Promise<string[]>
  /**
   * EIP-1474: Remote procedure call specification
   * https://eips.ethereum.org/EIPS/eip-1474
   */
  request(
    args: { method: 'web3_clientVersion' | 'klay_clientVersion' } | { method: 'klay_chainId' | 'eth_chainId' }
  ): Promise<string>

  /**
   *
   *
   */
  request(args: {
    method: 'klay_sign' | 'eth_sign'
    // params: [message: Bytes | string, account: Bytes | string, password: Bytes | string]
    params: [account: Bytes | string, message: Bytes | string]
  }): Promise<string>

  /**
   * EIP-2255: Wallet Permissions System
   * https://eips.ethereum.org/EIPS/eip-2255
   */
  request(
    args:
      | {
          method: 'wallet_requestPermissions'
          params: [{ klay_accounts: Record<string, any> }]
        }
      | {
          method: 'wallet_getPermissions'
        }
  ): Promise<WalletPermission[]>

  /**
   * EIP-3085: Wallet Add Ethereum Chain RPC Method
   * https://eips.ethereum.org/EIPS/eip-3085
   */

  /**
   * EIP-3326: Wallet Switch Ethereum Chain RPC Method
   * https://eips.ethereum.org/EIPS/eip-3326
   */
  request(
    args:
      | {
          method: 'wallet_addKlaytnChain'
          params: AddKlaytnChainParameter[]
        }
      | {
          method: 'wallet_addEthereumChain'
          params: AddKlaytnChainParameter[]
        }
      | {
          method: 'wallet_switchKlaytnChain'
          params: [{ chainId: string }]
        }
      | {
          method: 'wallet_switchEthereumChain'
          params: [{ chainId: string }]
        }
  ): Promise<null>
}

type InjectedProviderFlags = {
  isBitKeep?: true
  isBraveWallet?: true
  isCoinbaseWallet?: true
  isExodus?: true
  isFrame?: true
  isMathWallet?: true
  isMetaMask?: true
  isOneInchAndroidWallet?: true
  isOneInchIOSWallet?: true
  isOpera?: true
  isKaikas?: true
  isTokenPocket?: true
  isTokenary?: true
  isTrust?: true
}

type InjectedProviders = InjectedProviderFlags & {
  isMetaMask: true
  /** Only exists in MetaMask as of 2022/04/03 */
  _events: {
    connect?: () => void
  }
  /** Only exists in MetaMask as of 2022/04/03 */
  _state?: {
    accounts?: string[]
    initialized?: boolean
    isConnected?: boolean
    isPermanentlyDisconnected?: boolean
    isUnlocked?: boolean
  }
}

type WatchAssetParams = {
  /** In the future, other standards will be supported */
  type: 'ERC20'
  options: {
    /** Address of token contract */
    address: string
    /** Number of token decimals */
    decimals: number
    /** String url of token logo */
    image?: string
    /** A ticker symbol or shorthand, up to 5 characters */
    symbol: string
  }
}

type AddKlaytnChainParameter = {
  /** A 0x-prefixed hexadecimal string */
  chainId: string
  chainName: string
  nativeCurrency?: {
    name: string
    /** 2-6 characters long */
    symbol: string
    decimals: number
  }
  rpcUrls: string[]
  blockExplorerUrls?: string[]
  /** Currently ignored. */
  iconUrls?: string[]
}

type WalletPermissionCaveat = {
  type: string
  value: any
}

type WalletPermission = {
  caveats: WalletPermissionCaveat[]
  date: number
  id: string
  invoker: `http://${string}` | `https://${string}`
  parentCapability: 'klay_accounts' | string
}

export type InjectedConnectorOptions = {
  /** Name of connector */
  name?: string | ((detectedName: string | string[]) => string)
  /**
   * MetaMask 10.9.3 emits disconnect event when chain is changed.
   * This flag prevents the `"disconnect"` event from being emitted upon switching chains.
   * @see https://github.com/MetaMask/metamask-extension/issues/13375#issuecomment-1027663334
   */
  shimChainChangedDisconnect?: boolean
  /**
   * MetaMask and other injected providers do not support programmatic disconnect.
   * This flag simulates the disconnect behavior by keeping track of connection status in storage.
   * @see https://github.com/MetaMask/metamask-extension/issues/10353
   * @default true
   */
  shimDisconnect?: boolean
}

export type ConnectorOptions = Pick<InjectedConnectorOptions, 'shimChainChangedDisconnect' | 'shimDisconnect'> & {
  /**
   * While "disconnected" with `shimDisconnect`, allows user to select a different kaikas account (than the currently connected account) when trying to connect.
   */
  UNSTABLE_shimOnConnectSelectAccount?: boolean
}